requirements.txt
Flask
gunicorn
Pillow
torch
torchvision
numpy
opencv-python
scikit-learn